🔨refactor: separates tasks as individual modules
