+++ tasks/image_classification_timm/__init__.py
@@ -1,0 +1,1 @@
+from .task import *
