+++ tasks/question_answering_hf/__init__.py
@@ -1,0 +1,1 @@
+from .task import *
