+++ tasks/question_answering_hf/task.py
@@ -10,9 +10,9 @@ from transformers import (
     default_data_collator,
     BitsAndBytesConfig,
 )
-from .custom_model import prepare_custom_model_support
+from nyuntam_adapt.tasks.custom_model import prepare_custom_model_support
 from nyuntam_adapt.core.base_task import BaseTask
-from nyuntam_adapt.utils import prepare_model_for_kbit_training
+from nyuntam_adapt.utils.task_utils import prepare_model_for_kbit_training
 
 
 class ModelLoadingError(RuntimeError):
