+++ tasks/text_generation_hf/__init__.py
@@ -1,0 +1,1 @@
+from .task import *
