+++ tasks/translation_hf/task.py
@@ -1,6 +1,3 @@
-# input : document and label = summary
-
-import nltk
 import numpy as np
 from evaluate import load
 from transformers import (
@@ -11,11 +8,9 @@ from transformers import (
     Seq2SeqTrainingArguments,
     BitsAndBytesConfig,
 )
-from .custom_model import prepare_custom_model_support
-from nyuntam_adapt.utils import prepare_model_for_kbit_training
 from nyuntam_adapt.core.base_task import BaseTask
-
-nltk.download("punkt")
+from nyuntam_adapt.utils.task_utils import prepare_model_for_kbit_training
+from nyuntam_adapt.tasks.custom_model import prepare_custom_model_support
 
 
 class ModelLoadingError(RuntimeError):
@@ -24,15 +19,18 @@ class ModelLoadingError(RuntimeError):
     pass
 
 
-class Seq2Seq(BaseTask):
+class Translation(BaseTask):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.subtask = kwargs.get("subtask", None)
         self.eval_metric = kwargs.get("eval_metric", None)
-        self.max_input_length = kwargs.get("max_input_length", 512)
+        self.max_input_length = kwargs.get("max_input_length", 128)
+        self.source_lang = kwargs.get("source_lang", "en")
+        self.target_lang = kwargs.get("target_lang", "ro")
+        self.column_name = kwargs.get("FORMAT_NAMES", "translation")
         self.max_target_length = kwargs.get("max_target_length", 128)
-        self.input_column = kwargs["DATASET_ARGS"].get("input_column", "document")
-        self.target_column = kwargs["DATASET_ARGS"].get("target_column", "summary")
+        self.input_column = kwargs["DATASET_ARGS"].get("input_column", "english-text")
+        self.target_column = kwargs["DATASET_ARGS"].get("target_column", "roman-text")
         predict_with_generate = kwargs["TRAINING_ARGS"].get(
             "predict_with_generate", True
         )
@@ -44,13 +42,17 @@ class Seq2Seq(BaseTask):
             predict_with_generate=predict_with_generate,
             generation_max_length=generation_max_length,
         )
+        # if self.model_name in ["t5-small", "t5-base", "t5-larg", "t5-3b", "t5-11b"]:
+        #     self.prefix = "translate English to Romanian: "
+        # else:
+        #     self.prefix = ""
         self.prefix = kwargs.get("PREFIX", "")
         self.flash_attention = kwargs.get("flash_attention2", False)
         self.model_args = {}
 
     def prepare_model(self):
-        use_bnb = False
 
+        use_bnb = False
         if self.bnb_config["USE_4BIT"]["load_in_4bit"]:
             bnb_config_inputs = self.bnb_config["USE_4BIT"]
             use_bnb = True
@@ -65,7 +67,7 @@ class Seq2Seq(BaseTask):
         if self.local_model_path is not None:
             model, self.tokenizer, model_config = prepare_custom_model_support(
                 self.local_model_path,
-                "summarization",
+                "translation",
                 self.model_path,
                 num_gpu=self.num_gpu,
                 device=self.device,
@@ -73,6 +75,8 @@ class Seq2Seq(BaseTask):
                 use_flash_attention_2=self.flash_attention,
                 **self.model_args,
             )
+            # if use_bnb:
+            #     model = load_model_from_checkpoint(model,self.flash_attention,bnb_config)
 
         else:
             model_config = AutoConfig.from_pretrained(self.model_path)
@@ -113,73 +117,76 @@ class Seq2Seq(BaseTask):
             model = prepare_model_for_kbit_training(model, self.gradient_checkpointing)
         elif self.gradient_checkpointing:
             model.gradient_checkpointing_enable()
-
         self.collate_fn = DataCollatorForSeq2Seq(tokenizer=self.tokenizer)
 
         return model, self.tokenizer, model_config
 
     def preprocess_function(self, examples):
-        inputs = [self.prefix + doc for doc in examples[self.input_column]]
+        inputs = [
+            self.prefix + ex[self.source_lang] for ex in examples[self.column_name]
+        ]
+
+        targets = [ex[self.target_lang] for ex in examples[self.column_name]]
+
         model_inputs = self.tokenizer(
             inputs, max_length=self.max_input_length, truncation=True
         )
 
-        # Setup the tokenizer for targets
-        labels = self.tokenizer(
-            text_target=examples[self.target_column],
-            max_length=self.max_target_length,
-            truncation=True,
-        )
+        with self.tokenizer.as_target_tokenizer():
+            labels = self.tokenizer(
+                targets, max_length=self.max_target_length, truncation=True
+            )
 
         model_inputs["labels"] = labels["input_ids"]
+
         return model_inputs
 
-    def compute_metrics(self, p):
+    def postprocess(self, preds, labels):
+        preds = [pred.strip() for pred in preds]
+        labels = [[label.strip()] for label in labels]
+
+        return preds, labels
+
+    def compute_metrics(self, eval_preds):
         metric = load(self.eval_metric)
-        predictions, labels = p
+        predictions, labels = eval_preds
 
         predictions = np.where(
             predictions != -100, predictions, self.tokenizer.pad_token_id
         )
-
         decoded_preds = self.tokenizer.batch_decode(
             predictions, skip_special_tokens=True
         )
-        # Replace -100 in the labels as we can't decode them.
+
         labels = np.where(labels != -100, labels, self.tokenizer.pad_token_id)
-        decoded_labels = self.tokenizer.batch_decode(labels, skip_special_tokens=True)
-
-        # Rouge expects a newline after each sentence
-        decoded_preds = [
-            "\n".join(nltk.sent_tokenize(pred.strip())) for pred in decoded_preds
-        ]
-        decoded_labels = [
-            "\n".join(nltk.sent_tokenize(label.strip())) for label in decoded_labels
-        ]
-
-        # Note that other metrics may not have a `use_aggregator` parameter
-        # and thus will return a list, computing a metric for each sentence.
-        result = metric.compute(
-            predictions=decoded_preds,
-            references=decoded_labels,
-            use_stemmer=True,
-            use_aggregator=True,
+        decoded_labels = self.tokenizer.batch_decode(
+            labels, skip_special_tokens=True, clean_up_tokenization_spaces=True
         )
-        # Extract a few results
-        result = {key: value * 100 for key, value in result.items()}
 
-        # Add mean generated length
-        prediction_lens = [
-            np.count_nonzero(pred != self.tokenizer.pad_token_id)
-            for pred in predictions
-        ]
-        result["gen_len"] = np.mean(prediction_lens)
-
-        return {k: round(v, 4) for k, v in result.items()}
+        decoded_preds, decoded_labels = self.postprocess(decoded_preds, decoded_labels)
+        result = metric.compute(predictions=decoded_preds, references=decoded_labels)
+        # Additional result processing
+        if self.eval_metric == "sacrebleu":
+            result = {"sacrebleu": result["score"]}
+            prediction_lens = [
+                np.count_nonzero(pred != self.tokenizer.pad_token_id)
+                for pred in predictions
+            ]
+            result["gen_len"] = np.mean(prediction_lens)
+            return {k: round(v, 4) for k, v in result.items()}
+        else:
+            result = {key: value * 100 for key, value in result.items()}
+            prediction_lens = [
+                np.count_nonzero(pred != self.tokenizer.pad_token_id)
+                for pred in predictions
+            ]
+            result["gen_len"] = np.mean(prediction_lens)
+            return {k: round(v, 4) for k, v in result.items()}
 
     def prepare_dataset(self, dataset, processor):
         try:
             self.tokenizer = AutoTokenizer.from_pretrained(self.local_model_path)
         except:
             self.tokenizer = AutoTokenizer.from_pretrained(self.model_path)
+
         return dataset.map(self.preprocess_function, batched=True)
